test: route CDC status prints through output and silence in stress binary too
Move the quiet-output setup into an init() in a file tagged `integration ||
stress` so it also applies to the CDC stress binary (TestMain is integration-
tagged only), and route postgres_cdc / strategy status prints through the
internal/output seam so output.Init(io.Discard) silences them.
