+++ pkg/source/postgres_cdc/source.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"github.com/bruin-data/ingestr/internal/config"
+	"github.com/bruin-data/ingestr/internal/output"
 	"github.com/bruin-data/ingestr/pkg/source"
 	"github.com/jackc/pglogrepl"
 	"github.com/jackc/pgx/v5"
@@ -1002,7 +1003,7 @@ func ensureManagedPublication(ctx context.Context, pool *pgxpool.Pool, pubName s
 	}
 
 	for _, s := range skipped {
-		fmt.Printf("Warning: %s\n", s.warning(pubName))
+		output.Statusf("Warning: %s\n", s.warning(pubName))
 	}
 
 	tableList := quotePublicationTables(included)
@@ -1234,7 +1235,7 @@ func (s *PostgresCDCSource) warnKeylessTable(name string) {
 		return
 	}
 	s.keylessWarned[name] = true
-	fmt.Printf("Warning: table %s has no primary key or replica identity index; ingesting it as an append-only change log (_cdc_deleted marks deletes, updates arrive as delete+insert pairs)\n", name)
+	output.Statusf("Warning: table %s has no primary key or replica identity index; ingesting it as an append-only change log (_cdc_deleted marks deletes, updates arrive as delete+insert pairs)\n", name)
 }
 
 // publicationTableFullName renders a table name the way GetTables and the WAL
