+++ pkg/strategy/strategy.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/bruin-data/ingestr/internal/annotation"
 	"github.com/bruin-data/ingestr/internal/config"
+	"github.com/bruin-data/ingestr/internal/output"
 	"github.com/bruin-data/ingestr/pkg/destination"
 	"github.com/bruin-data/ingestr/pkg/progress"
 	"github.com/bruin-data/ingestr/pkg/schema"
@@ -251,7 +252,7 @@ func applyEvolutionPlanIfIncarnation(ctx context.Context, dest destination.Desti
 		return "", err
 	}
 	for _, w := range warnings {
-		fmt.Printf("Warning: %s\n", w)
+		output.Statusf("Warning: %s\n", w)
 	}
 	// Mark the plan applied so a repeat call is a no-op. This mirrors the prior
 	// EvolutionPlan.Apply contract (which cleared its rendered statements) and
